refactor(server): extract Express app builder to app.ts and slim server.ts
